# docs/api/ui-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Issue Orchestrator UI API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/issues/{issue_number}/timeline-evidence/pin": {
      "put": {
        "summary": "Set the pin state for one exact Timeline run",
        "parameters": [
          {
            "name": "issue_number",
            "in": "path",
            "required": true,
            "schema": {"type": "integer", "minimum": 1}
          },
          {
            "name": "repo_root",
            "in": "query",
            "required": false,
            "description": "Control Center repository scope; omitted on a Repository Engine.",
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/TimelineEvidencePinRequestPayload"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "The resulting exact-run evidence state",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TimelineEvidenceStatePayload"}
              }
            }
          },
          "404": {"description": "The exact run or its evidence is unavailable"},
          "409": {"description": "The exact issue/run pair is not eligible for pinning"}
        }
      }
    },
...
    "/api/issue-detail/{issue_number}": {
      "get": {
        "summary": "Get issue detail payload",
        "parameters": [
          {
            "name": "issue_number",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {"$ref": "#/components/schemas/TimelineView", "default": "user"}
          },
          {
            "name": "repo_root",
            "in": "query",
            "required": false,
            "description": "Control Center repository scope; omitted on a Repository Engine.",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Issue detail",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/IssueDetailPayload"}
              }
            }
          }
        }
      }
    },
    "/api/session/terminal-recording/{issue_number}": {
      "get": {
        "summary": "Read an exact run's terminal recording while evidence is retained",
        "parameters": [
          {
            "name": "issue_number",
            "in": "path",
            "required": true,
            "schema": {"type": "integer", "minimum": 1}
          },
          {
            "name": "run_dir",
            "in": "query",
            "required": true,
            "schema": {"type": "string"}
          },
          {
            "name": "repo_root",
            "in": "query",
            "required": false,
            "description": "Control Center repository scope; omitted on a Repository Engine.",
            "schema": {"type": "string"}
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 0}
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 200}
          },
          {
            "name": "round_index",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "minimum": 1}
          },
          {
            "name": "session_role",
            "in": "query",
            "required": false,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Terminal recording replay payload",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TerminalRecordingPayload"}
              }
            }
          },
          "410": {"description": "The exact run's Timeline evidence expired or is unavailable"}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IssueDetailValidationDiagnosticPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "state",
          "run_dir",
          "session_name",
          "reason",
          "suite",
          "command",
          "exit_code",
          "failed_tests",
          "failed_tests_preview",
          "validation_record_path",
          "validation_stderr",
          "validation_stdout",
          "junit_cases"
        ],
        "properties": {
          "state": {"type": "string"},
          "run_dir": {"type": "string"},
          "session_name": {"type": ["string", "null"]},
          "reason": {"type": "string"},
          "suite": {"type": "string"},
          "command": {"type": "string"},
          "exit_code": {"type": "integer"},
          "failed_tests": {"type": "array", "items": {"type": "string"}},
          "failed_tests_preview": {"type": "array", "items": {"type": "string"}},
          "validation_record_path": {"type": ["string", "null"]},
          "validation_stderr": {"type": ["string", "null"]},
          "validation_stdout": {"type": ["string", "null"]},
          "junit_cases": {"type": "array", "items": {"$ref": "#/components/schemas/TestCaseResultPayload"}}
        }
      },
      "IssueDetailTimelineDiagnosticPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "state",
          "signals",
          "expected_timeline_store",
          "expected_timeline_store_exists",
          "resolved_run_dir",
          "dropped_missing_semantics"
        ],
        "properties": {
          "state": {"type": "string"},
          "signals": {"type": "array", "items": {"type": "string"}},
          "expected_timeline_store": {"type": "string"},
          "expected_timeline_store_exists": {"type": "boolean"},
          "resolved_run_dir": {"type": ["string", "null"]},
          "dropped_missing_semantics": {"type": "integer"}
        }
      },
      "IssueDetailSummaryPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "last_event", "event_count"],
        "properties": {
          "status": {"type": "string"},
          "last_event": {"type": "string"},
          "event_count": {"type": "integer"},
          "run_diagnostic": {
            "oneOf": [
              {"$ref": "#/components/schemas/IssueDetailValidationDiagnosticPayload"},
              {"type": "null"}
            ]
          },
          "timeline_diagnostic": {
            "oneOf": [
              {"$ref": "#/components/schemas/IssueDetailTimelineDiagnosticPayload"},
              {"type": "null"}
            ]
          }
        }
      },
      "IssueDetailActionPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "label"],
        "properties": {
          "id": {"type": "string"},
          "label": {"type": "string"},
          "url": {"type": ["string", "null"]},
          "run_dir": {"type": ["string", "null"]}
        }
      },
      "IssueDetailBlockedDetailPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reason", "labels", "rework_info", "event_summary"],
        "properties": {
          "reason": {"type": "string"},
          "labels": {"type": "array", "items": {"type": "string"}},
          "rework_info": {"type": ["string", "null"]},
          "event_summary": {"type": "string"}
        }
      },
      "StackDependencyGatePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["gate", "open", "reason_codes", "reasons"],
        "properties": {
          "gate": {"type": "string"},
          "open": {"type": "boolean"},
          "reason_codes": {"type": "array", "items": {"type": "string"}},
          "reasons": {"type": "array", "items": {"type": "string"}}
        }
      },
      "StackDependencyPredecessorPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ref", "mode", "state", "problem"],
        "properties": {
          "ref": {"type": "string"},
          "mode": {"type": "string"},
          "state": {"type": "string"},
          "problem": {"type": ["string", "null"]}
        }
      },
      "StackDependencySuccessorPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["issue_number", "ref", "mode"],
        "properties": {
          "issue_number": {"type": "integer"},
          "ref": {"type": "string"},
          "mode": {"type": "string"}
        }
      },
      "StackDependencyGateViewPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "issue_number",
          "mode",
          "has_stack_edges",
          "gates",
          "predecessors",
          "successors",
          "blocked_gates",
          "blocked_reason_codes",
          "stale",
          "stale_reason_codes",
          "stack_base_branch",
          "approval_freshness"
        ],
        "properties": {
          "issue_number": {"type": "integer"},
          "mode": {"type": "string"},
          "has_stack_edges": {"type": "boolean"},
          "gates": {"type": "array", "items": {"$ref": "#/components/schemas/StackDependencyGatePayload"}},
          "predecessors": {"type": "array", "items": {"$ref": "#/components/schemas/StackDependencyPredecessorPayload"}},
          "successors": {"type": "array", "items": {"$ref": "#/components/schemas/StackDependencySuccessorPayload"}},
          "blocked_gates": {"type": "array", "items": {"type": "string"}},
          "blocked_reason_codes": {"type": "array", "items": {"type": "string"}},
          "stale": {"type": "boolean"},
          "stale_reason_codes": {"type": "array", "items": {"type": "string"}},
          "stack_base_branch": {"type": ["string", "null"]},
          "approval_freshness": {"type": "string"}
        }
      },
      "IssueDetailPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "issue_number",
          "title",
          "issue_url",
          "phase_toc",
          "cycles",
          "events",
          "summary",
          "actions",
          "status_explanation",
          "runs",
          "timeline_steps",
          "run_count",
          "previous_runs",
          "previous_runs_count",
          "raw_events_count",
          "blocked_detail"
        ],
        "properties": {
          "issue_number": {"type": "integer"},
          "e2e_run_id": {"type": ["integer", "null"]},
          "title": {"type": "string"},
          "issue_url": {"type": "string"},
          "phase_toc": {"type": "array", "items": {"type": "object", "additionalProperties": true}},
          "cycles": {"type": "array", "items": {"type": "object", "additionalProperties": true}},
          "events": {"type": "array", "items": {"type": "object", "additionalProperties": true}},
          "summary": {"$ref": "#/components/schemas/IssueDetailSummaryPayload"},
          "actions": {"type": "array", "items": {"$ref": "#/components/schemas/IssueDetailActionPayload"}},
          "status_explanation": {"type": "string"},
          "runs": {"type": "array", "items": {"$ref": "#/components/schemas/JourneyRunPayload"}},
          "timeline_steps": {"type": "array", "items": {"type": "object", "additionalProperties": true}},
          "run_count": {"type": "integer"},
          "previous_runs": {"type": "array", "items": {"type": "object", "additionalProperties": true}},
          "previous_runs_count": {"type": "integer"},
          "raw_events_count": {"type": "integer"},
          "blocked_detail": {
            "oneOf": [
              {"$ref": "#/components/schemas/IssueDetailBlockedDetailPayload"},
              {"type": "null"}
            ]
          },
          "stack_dependency": {
            "oneOf": [
              {"$ref": "#/components/schemas/StackDependencyGateViewPayload"},
              {"type": "null"}
            ]
          },
          "lifecycle": {
            "oneOf": [
              {"$ref": "#/components/schemas/LifecycleTimelineContainerPayload"},
              {"type": "null"}
            ]
          },
          "view": {"$ref": "#/components/schemas/TimelineView", "default": "user"}
        }
      },
      "TimelineEvidencePinRequestPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["run_dir", "pinned"],
        "properties": {
          "run_dir": {"type": "string", "minLength": 1},
          "pinned": {"type": "boolean"}
        }
      },
      "TimelineEvidenceStatePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "run_dir",
          "status",
          "label",
          "available",
          "pinned",
          "archived",
          "expires_at",
          "help_text",
          "unpin_expires_immediately"
        ],
        "properties": {
          "run_dir": {"type": "string"},
          "status": {"type": "string", "enum": ["active", "retained", "pinned", "expired", "missing"]},
          "label": {"type": "string"},
          "available": {"type": "boolean"},
          "pinned": {"type": "boolean"},
          "archived": {"type": "boolean"},
          "expires_at": {"type": ["string", "null"]},
          "help_text": {"type": "string"},
          "unpin_expires_immediately": {"type": "boolean"}
        }
      },
      "TerminalRecordingPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["issue_number", "recording_path", "render_mode"],
        "properties": {
          "issue_number": {"type": "integer", "minimum": 1},
          "recording_path": {"type": "string"},
          "render_mode": {"type": "string", "enum": ["terminal", "transcript"]},
          "transcript_lines": {
            "type": ["array", "null"],
            "items": {"type": "string"}
          },
          "transcript_hash": {"type": ["string", "null"]},
          "unchanged": {"type": "boolean"},
          "content_type": {"type": "string"},
          "total_events": {"type": "integer", "minimum": 0},
          "initial_geometry": {
            "type": ["object", "null"],
            "additionalProperties": {"type": "integer"}
          },
          "offset": {"type": "integer", "minimum": 0},
          "truncated": {"type": "boolean"},
          "chapters": {
            "type": ["array", "null"],
            "items": {"type": "object", "additionalProperties": true}
          },
          "recording_event_index": {"type": ["integer", "null"], "minimum": 0},
          "events": {
            "type": "array",
            "items": {"type": "object", "additionalProperties": true}
          }
        }
      },
...
      "TestCaseIssueLinkPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["number", "status", "resolution"],
        "properties": {
          "number": {"type": "integer"},
          "status": {"type": "string"},
          "resolution": {"type": ["string", "null"]}
        }
      },
      "TestCaseHistoryPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["outcome", "run_id"],
        "properties": {
          "outcome": {"type": "string"},
          "run_id": {"type": "integer"}
        }
      },
      "CapturedOutputAvailabilityPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["stdout_available", "stderr_available"],
        "properties": {
          "stdout_available": {"type": "boolean"},
          "stderr_available": {"type": "boolean"}
        }
      },
      "TestCaseResultPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "nodeid",
          "case_id",
          "label",
          "display_name",
          "suite_name",
          "result_source",
          "outcome",
          "duration_seconds",
          "longrepr",
          "failure_summary",
          "retry_outcome",
          "is_quarantined",
          "updated_at",
          "history",
          "existing_issue",
          "category",
          "result_category",
          "flip_rate",
          "flip_rate_percent",
          "is_likely_flaky",
          "captured_output"
        ],
        "properties": {
          "nodeid": {"type": "string"},
          "case_id": {"type": "string"},
          "label": {"type": "string"},
          "display_name": {"type": ["string", "null"]},
          "suite_name": {"type": ["string", "null"]},
          "result_source": {"type": "string"},
          "outcome": {"type": "string"},
          "duration_seconds": {"type": ["number", "null"]},
          "longrepr": {"type": ["string", "null"]},
          "failure_summary": {"type": ["string", "null"]},
          "retry_outcome": {"type": ["string", "null"]},
          "is_quarantined": {"type": "boolean"},
          "updated_at": {"type": "string"},
          "history": {"type": "array", "items": {"$ref": "#/components/schemas/TestCaseHistoryPayload"}},
          "existing_issue": {
            "oneOf": [
              {"$ref": "#/components/schemas/TestCaseIssueLinkPayload"},
              {"type": "null"}
            ]
          },
          "category": {"type": "string"},
          "result_category": {"type": "string"},
          "flip_rate": {"type": "number"},
          "flip_rate_percent": {"type": "number"},
          "is_likely_flaky": {"type": "boolean"},
          "captured_output": {"$ref": "#/components/schemas/CapturedOutputAvailabilityPayload"}
        }
      },
...
      "TimelineDiagnosticPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "severity"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "severity": {"enum": ["info", "warning", "error"]},
          "evidence_ref": {"type": ["string", "null"]}
        }
      },
...
      "TimelineSubjectPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "id", "label"],
        "properties": {
          "kind": {"enum": ["dashboard", "issue", "e2e_suite", "e2e_run"]},
          "id": {"type": "string"},
          "label": {"type": "string"},
          "status": {"type": ["string", "null"]},
          "outcome": {"type": ["string", "null"]}
        }
      },
...
      "TimelineView": {
        "type": "string",
        "enum": ["user", "ops", "debug", "raw"],
        "description": "Which timeline lens the UI renders: user (story), ops, debug, or raw events. Reused by every issue-detail / E2E-run timeline query param and payload field so the wire values cannot drift."
      },
...
      "RunningCodingAttemptPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "issue_number", "agent", "started_at", "session_recording", "commands"],
        "properties": {
          "kind": {"const": "running_coding_attempt"},
          "issue_number": {"type": "integer"},
          "agent": {"$ref": "#/components/schemas/AgentIdentityPayload"},
          "started_at": {"type": "string"},
          "session_recording": {"$ref": "#/components/schemas/SessionRecordingEvidencePayload"},
          "commands": {"type": "array", "items": {"$ref": "#/components/schemas/TimelineCommandPayload"}}
        }
      },
...
      "BlockedCodingAttemptPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "kind", "issue_number", "agent", "blocked_at", "reason",
          "session_recording", "diagnostics", "commands"
        ],
        "properties": {
          "kind": {"const": "blocked_coding_attempt"},
          "issue_number": {"type": "integer"},
          "agent": {"$ref": "#/components/schemas/AgentIdentityPayload"},
          "started_at": {"type": ["string", "null"]},
          "blocked_at": {"type": "string"},
          "reason": {"type": "string"},
          "session_recording": {"$ref": "#/components/schemas/SessionRecordingEvidencePayload"},
          "diagnostics": {"type": "array", "items": {"$ref": "#/components/schemas/TimelineDiagnosticPayload"}},
          "commands": {"type": "array", "items": {"$ref": "#/components/schemas/TimelineCommandPayload"}}
        }
      },
...
      "CodingAttemptPayload": {
        "oneOf": [
          {"$ref": "#/components/schemas/RunningCodingAttemptPayload"},
          {"$ref": "#/components/schemas/CompletedCodingAttemptPayload"},
          {"$ref": "#/components/schemas/PublishFailedCodingAttemptPayload"},
          {"$ref": "#/components/schemas/BlockedCodingAttemptPayload"},
          {"$ref": "#/components/schemas/FailedCodingAttemptPayload"},
          {"$ref": "#/components/schemas/MissingCodingEvidencePayload"}
        ],
        "discriminator": {"propertyName": "kind"}
      },
...
      "ReviewStagePayload": {
        "oneOf": [
          {"$ref": "#/components/schemas/ReviewNotReachedPayload"},
          {"$ref": "#/components/schemas/ReviewSkippedPayload"},
          {"$ref": "#/components/schemas/ReviewRunningPayload"},
          {"$ref": "#/components/schemas/ReviewApprovedPayload"},
          {"$ref": "#/components/schemas/ReviewChangesRequestedPayload"},
          {"$ref": "#/components/schemas/ReviewFailedPayload"},
          {"$ref": "#/components/schemas/MissingReviewEvidencePayload"}
        ],
        "discriminator": {"propertyName": "kind"}
      },
      "CycleArtifactsPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "log_url", "pr_url", "pr_number", "has_review_feedback",
          "run_dir", "review_report", "review_decision"
        ],
        "properties": {
          "log_url": {"type": ["string", "null"]},
          "pr_url": {"type": ["string", "null"]},
          "pr_number": {"type": ["integer", "null"]},
          "has_review_feedback": {"type": "boolean"},
          "run_dir": {"type": ["string", "null"]},
          "review_report": {
            "oneOf": [
              {"$ref": "#/components/schemas/CycleReviewArtifactPayload"},
              {"type": "null"}
            ]
          },
          "review_decision": {
            "oneOf": [
              {"$ref": "#/components/schemas/CycleReviewArtifactPayload"},
              {"type": "null"}
            ]
          }
        }
      },
...
      "JourneyStepPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["timestamp", "time_label", "day", "narrative", "status", "event", "actions"],
        "properties": {
          "timestamp": {"type": "string"},
          "time_label": {"type": "string"},
          "day": {"type": "string"},
          "narrative": {"type": "string"},
          "status": {"type": "string"},
          "event": {"type": "string"},
          "detail": {"type": ["string", "null"]},
          "actions": {"type": "array", "items": {"type": "object", "additionalProperties": true}},
          "evidence": {
            "oneOf": [
              {"$ref": "#/components/schemas/TimelineEvidenceStatePayload"},
              {"type": "null"}
            ]
          },
          "in_round_progress": {"type": "boolean", "default": false}
        }
      },
      "JourneyPhaseGroupPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["key", "label", "steps"],
        "properties": {
          "key": {"enum": ["coding", "review", "rework", "orchestrator"]},
          "label": {"type": "string"},
          "steps": {"type": "array", "items": {"$ref": "#/components/schemas/JourneyStepPayload"}}
        }
      },
      "CycleValidationBadgePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["state", "command"],
        "properties": {
          "state": {"enum": ["pending", "not_validated", "passed", "failed"]},
          "command": {
            "oneOf": [
              {"$ref": "#/components/schemas/OpenValidationDetailsCommandPayload"},
              {"type": "null"}
            ]
          }
        }
      },
      "OutcomeBadgePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["label", "tone"],
        "properties": {
          "label": {"type": "string"},
          "tone": {"enum": ["passed", "failed", "error", "in_progress", "neutral"]}
        }
      },
      "IssueCyclePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "cycle_number", "coder", "review", "outcome", "diagnostics",
          "lifecycle", "iteration", "run_id", "timestamp", "session_run_ids",
          "agent", "reviewer_agent", "retry_count", "reset_from_scratch",
          "cycle_label", "time_label", "expanded", "cycle_in_run",
          "artifacts", "steps", "phase_groups", "validation"
        ],
        "properties": {
          "cycle_number": {"type": "integer"},
          "coder": {"$ref": "#/components/schemas/CodingAttemptPayload"},
          "review": {"$ref": "#/components/schemas/ReviewStagePayload"},
          "outcome": {"$ref": "#/components/schemas/OutcomeBadgePayload"},
          "diagnostics": {"type": "array", "items": {"$ref": "#/components/schemas/TimelineDiagnosticPayload"}},
          "lifecycle": {"type": ["integer", "null"]},
          "iteration": {"type": ["integer", "null"]},
          "run_id": {"type": ["string", "null"]},
          "timestamp": {"type": ["string", "null"]},
          "session_run_ids": {"type": "array", "items": {"type": "string"}},
          "agent": {"type": ["string", "null"]},
          "reviewer_agent": {"type": ["string", "null"]},
          "retry_count": {"type": ["integer", "null"]},
          "reset_from_scratch": {"type": ["boolean", "null"]},
          "cycle_label": {"type": ["string", "null"]},
          "time_label": {"type": ["string", "null"]},
          "expanded": {"type": ["boolean", "null"]},
          "cycle_in_run": {"type": ["integer", "null"]},
          "artifacts": {
            "oneOf": [
              {"$ref": "#/components/schemas/CycleArtifactsPayload"},
              {"type": "null"}
            ]
          },
          "steps": {"type": "array", "items": {"$ref": "#/components/schemas/JourneyStepPayload"}},
          "phase_groups": {"type": "array", "items": {"$ref": "#/components/schemas/JourneyPhaseGroupPayload"}},
          "validation": {
            "oneOf": [
              {"$ref": "#/components/schemas/CycleValidationBadgePayload"},
              {"type": "null"}
            ]
          }
        }
      },
      "JourneyRunPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "run_number", "run_label", "outcome", "run_key", "run_id",
          "session_run_ids", "timestamp", "time_label", "expanded",
          "reset_from_scratch", "cycles"
        ],
        "properties": {
          "run_number": {"type": "integer"},
          "run_label": {"type": "string"},
          "outcome": {"$ref": "#/components/schemas/OutcomeBadgePayload"},
          "run_key": {"type": "string"},
          "run_id": {"type": ["string", "null"]},
          "session_run_ids": {"type": "array", "items": {"type": "string"}},
          "timestamp": {"type": "string"},
          "time_label": {"type": "string"},
          "expanded": {"type": "boolean"},
          "reset_from_scratch": {"type": "boolean"},
          "cycles": {"type": "array", "items": {"$ref": "#/components/schemas/IssueCyclePayload"}}
        }
      },
...
      "DashboardIterationPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "subject", "issue_lifecycles", "diagnostics"],
        "properties": {
          "kind": {"const": "dashboard_current"},
          "subject": {"$ref": "#/components/schemas/TimelineSubjectPayload"},
          "issue_lifecycles": {"type": "array", "items": {"$ref": "#/components/schemas/IssueLifecyclePayload"}},
          "diagnostics": {"type": "array", "items": {"$ref": "#/components/schemas/TimelineDiagnosticPayload"}}
        }
      },
      "DashboardTimelineContainerPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "subject", "current"],
        "properties": {
          "kind": {"const": "dashboard"},
          "subject": {"$ref": "#/components/schemas/TimelineSubjectPayload"},
          "current": {"$ref": "#/components/schemas/DashboardIterationPayload"}
        }
      },
...
      "E2ERunIterationPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "subject", "e2e_run", "diagnostics"],
        "properties": {
          "kind": {"const": "e2e_run"},
          "subject": {"$ref": "#/components/schemas/TimelineSubjectPayload"},
          "e2e_run": {"$ref": "#/components/schemas/E2ERunLifecyclePayload"},
          "diagnostics": {"type": "array", "items": {"$ref": "#/components/schemas/TimelineDiagnosticPayload"}}
        }
      },
      "E2ESuiteTimelineContainerPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "subject", "runs"],
        "properties": {
          "kind": {"const": "e2e_suite"},
          "subject": {"$ref": "#/components/schemas/TimelineSubjectPayload"},
          "runs": {"type": "array", "items": {"$ref": "#/components/schemas/E2ERunIterationPayload"}}
        }
      },
      "LifecycleTimelineContainerPayload": {
        "oneOf": [
          {"$ref": "#/components/schemas/DashboardTimelineContainerPayload"},
          {"$ref": "#/components/schemas/E2ESuiteTimelineContainerPayload"}
        ],
        "discriminator": {"propertyName": "kind"}
      },
...
    }
  }
}